rename `*Hoc` to `with*`
