+++ src/derived/root.js
@@ -2,7 +2,7 @@
 import * as React from 'react'
 import Grid from '../grid'
 import styles from './root.css'
-import type { Props } from '../base.hoc'
+import type { Props } from '../withBase'
 
 export default function Root({ children, justify, ...props }: Props) {
   return (
