+++ src/reflex.js
@@ -1,8 +1,8 @@
 // @flow
 import * as srcUtils from './utils'
 
-export { default as GridHoc } from './grid.hoc'
-export { default as BaseHoc } from './base.hoc'
+export { default as withCol } from './withGrid'
+export { default as withBase } from './withBase'
 export { default as Base } from './base'
 export { default as Grid } from './grid'
 export { default as Layout } from './layout'
